@@ -8,7 +8,6 @@ classifiers =
 	Operating System :: OS Independent
 	Programming Language :: Python
 	Programming Language :: Python :: 3
-	Programming Language :: Python :: 3.8
 	Programming Language :: Python :: 3.9
 	Programming Language :: Python :: 3.10
 	Programming Language :: Python :: 3.11
@@ -29,7 +28,7 @@ version = attr: pathspec._meta.__version__
 
 [options]
 packages = find:
-python_requires = >=3.8
+python_requires = >=3.9
 setup_requires = setuptools>=40.8.0
 test_suite = tests
 
