@@ -9,12 +9,13 @@ import pathlib
 import shutil
 import tempfile
 import unittest
+from collections.abc import (
+	Iterable)
 from functools import (
 	partial)
 from typing import (
-	Iterable,  # Replaced by `collections.abc.Iterable` in 3.9.
-	Optional,  # Replaced by `X | None` in 3.10.
-	Tuple)  # Replaced by `tuple` in 3.9.
+	ClassVar,
+	Optional)  # Replaced by `X | None` in 3.10.
 
 from pathspec.patterns.gitwildmatch import (
 	GitWildMatchPattern)
@@ -26,6 +27,7 @@ from pathspec.util import (
 	match_file,
 	normalize_file)
 from tests.util import (
+	get_paths_from_entries,
 	make_dirs,
 	make_files,
 	make_links,
@@ -115,6 +117,11 @@ class IterTreeTest(unittest.TestCase):
 	:meth:`.iter_tree_files` functions.
 	"""
 
+	no_symlink: ClassVar[bool]
+	"""
+	*no_symlink* (:class:`bool`) is whether symlinks are not supported.
+	"""
+
 	def make_dirs(self, dirs: Iterable[str]) -> None:
 		"""
 		Create the specified directories.
@@ -127,7 +134,7 @@ class IterTreeTest(unittest.TestCase):
 		"""
 		make_files(self.temp_dir, files)
 
-	def make_links(self, links: Iterable[Tuple[str, str]]) -> None:
+	def make_links(self, links: Iterable[tuple[str, str]]) -> None:
 		"""
 		Create the specified links.
 		"""
@@ -152,7 +159,37 @@ class IterTreeTest(unittest.TestCase):
 		"""
 		shutil.rmtree(self.temp_dir)
 
-	def test_01_files(self):
+	def test_01_files_1_entries(self):
+		"""
+		Tests to make sure all files are found.
+		"""
+		self.make_dirs([
+			'Empty',
+			'Dir',
+			'Dir/Inner',
+		])
+		self.make_files([
+			'a',
+			'b',
+			'Dir/c',
+			'Dir/d',
+			'Dir/Inner/e',
+			'Dir/Inner/f',
+		])
+		results = get_paths_from_entries(iter_tree_entries(self.temp_dir))
+		self.assertEqual(results, set(map(ospath, [
+			'a',
+			'b',
+			'Dir',
+			'Dir/Inner',
+			'Dir/Inner/e',
+			'Dir/Inner/f',
+			'Dir/c',
+			'Dir/d',
+			'Empty',
+		])))
+
+	def test_01_files_2_files(self):
 		"""
 		Tests to make sure all files are found.
 		"""
@@ -179,12 +216,11 @@ class IterTreeTest(unittest.TestCase):
 			'Dir/Inner/f',
 		])))
 
-	def test_02_link_1_check_1_symlink(self):
+	def test_02_link_1_check_symlink(self):
 		"""
 		Tests whether links can be created.
 		"""
-		# NOTE: Windows Vista and greater supports `os.symlink` for Python
-		# 3.2+.
+		# NOTE: Windows Vista and greater supports `os.symlink` for Python 3.2+.
 		no_symlink: Optional[bool] = None
 		try:
 			file = self.temp_dir / 'file'
@@ -201,7 +237,46 @@ class IterTreeTest(unittest.TestCase):
 		finally:
 			self.__class__.no_symlink = no_symlink
 
-	def test_02_link_2_links(self):
+	def test_02_link_2_links_1_entries(self):
+		"""
+		Tests to make sure links to directories and files work.
+		"""
+		self.require_symlink()
+		self.make_dirs([
+			'Dir',
+		])
+		self.make_files([
+			'a',
+			'b',
+			'Dir/c',
+			'Dir/d',
+		])
+		self.make_links([
+			('ax', 'a'),
+			('bx', 'b'),
+			('Dir/cx', 'Dir/c'),
+			('Dir/dx', 'Dir/d'),
+			('DirX', 'Dir'),
+		])
+		results = get_paths_from_entries(iter_tree_entries(self.temp_dir))
+		self.assertEqual(results, set(map(ospath, [
+			'a',
+			'ax',
+			'b',
+			'bx',
+			'Dir',
+			'Dir/c',
+			'Dir/cx',
+			'Dir/d',
+			'Dir/dx',
+			'DirX',
+			'DirX/c',
+			'DirX/cx',
+			'DirX/d',
+			'DirX/dx',
+		])))
+
+	def test_02_link_2_links_2_files(self):
 		"""
 		Tests to make sure links to directories and files work.
 		"""
@@ -238,7 +313,57 @@ class IterTreeTest(unittest.TestCase):
 			'DirX/dx',
 		])))
 
-	def test_02_link_3_sideways_links(self):
+	def test_02_link_3_sideways_links_1_entries(self):
+		"""
+		Tests to make sure the same directory can be encountered multiple
+		times via links.
+		"""
+		self.require_symlink()
+		self.make_dirs([
+			'Dir',
+			'Dir/Target',
+		])
+		self.make_files([
+			'Dir/Target/file',
+		])
+		self.make_links([
+			('Ax', 'Dir'),
+			('Bx', 'Dir'),
+			('Cx', 'Dir/Target'),
+			('Dx', 'Dir/Target'),
+			('Dir/Ex', 'Dir/Target'),
+			('Dir/Fx', 'Dir/Target'),
+		])
+		results = get_paths_from_entries(iter_tree_entries(self.temp_dir))
+		self.assertEqual(results, set(map(ospath, [
+			'Ax',
+			'Ax/Ex',
+			'Ax/Ex/file',
+			'Ax/Fx',
+			'Ax/Fx/file',
+			'Ax/Target',
+			'Ax/Target/file',
+			'Bx',
+			'Bx/Ex',
+			'Bx/Ex/file',
+			'Bx/Fx',
+			'Bx/Fx/file',
+			'Bx/Target',
+			'Bx/Target/file',
+			'Cx',
+			'Cx/file',
+			'Dx',
+			'Dx/file',
+			'Dir',
+			'Dir/Ex',
+			'Dir/Ex/file',
+			'Dir/Fx',
+			'Dir/Fx/file',
+			'Dir/Target',
+			'Dir/Target/file',
+		])))
+
+	def test_02_link_3_sideways_links_2_files(self):
 		"""
 		Tests to make sure the same directory can be encountered multiple
 		times via links.
@@ -274,7 +399,27 @@ class IterTreeTest(unittest.TestCase):
 			'Dir/Target/file',
 		])))
 
-	def test_02_link_4_recursive_links(self):
+	def test_02_link_4_recursive_links_1_entries(self):
+		"""
+		Tests detection of recursive links.
+		"""
+		self.require_symlink()
+		self.make_dirs([
+			'Dir',
+		])
+		self.make_files([
+			'Dir/file',
+		])
+		self.make_links([
+			('Dir/Self', 'Dir'),
+		])
+		with self.assertRaises(RecursionError) as context:
+			set(iter_tree_entries(self.temp_dir))
+
+		self.assertEqual(context.exception.first_path, 'Dir')
+		self.assertEqual(context.exception.second_path, ospath('Dir/Self'))
+
+	def test_02_link_4_recursive_links_2_files(self):
 		"""
 		Tests detection of recursive links.
 		"""
@@ -294,7 +439,37 @@ class IterTreeTest(unittest.TestCase):
 		self.assertEqual(context.exception.first_path, 'Dir')
 		self.assertEqual(context.exception.second_path, ospath('Dir/Self'))
 
-	def test_02_link_5_recursive_circular_links(self):
+	def test_02_link_5_recursive_circular_links_1_entries(self):
+		"""
+		Tests detection of recursion through circular links.
+		"""
+		self.require_symlink()
+		self.make_dirs([
+			'A',
+			'B',
+			'C',
+		])
+		self.make_files([
+			'A/d',
+			'B/e',
+			'C/f',
+		])
+		self.make_links([
+			('A/Bx', 'B'),
+			('B/Cx', 'C'),
+			('C/Ax', 'A'),
+		])
+		with self.assertRaises(RecursionError) as context:
+			set(iter_tree_entries(self.temp_dir))
+
+		self.assertIn(context.exception.first_path, ('A', 'B', 'C'))
+		self.assertEqual(context.exception.second_path, {
+			'A': ospath('A/Bx/Cx/Ax'),
+			'B': ospath('B/Cx/Ax/Bx'),
+			'C': ospath('C/Ax/Bx/Cx'),
+		}[context.exception.first_path])
+
+	def test_02_link_5_recursive_circular_links_2_files(self):
 		"""
 		Tests detection of recursion through circular links.
 		"""
@@ -324,7 +499,7 @@ class IterTreeTest(unittest.TestCase):
 			'C': ospath('C/Ax/Bx/Cx'),
 		}[context.exception.first_path])
 
-	def test_02_link_6_detect_broken_links(self):
+	def test_02_link_6_detect_broken_links_1_entries(self):
 		"""
 		Tests that broken links are detected.
 		"""
@@ -336,11 +511,36 @@ class IterTreeTest(unittest.TestCase):
 			('A', 'DOES_NOT_EXIST'),
 		])
 		with self.assertRaises(OSError) as context:
-			set(iter_tree_files(self.temp_dir, on_error=reraise))
+			set(iter_tree_entries(self.temp_dir, on_error=reraise))
 
 		self.assertEqual(context.exception.errno, errno.ENOENT)
 
-	def test_02_link_7_ignore_broken_links(self):
+	def test_02_link_6_detect_broken_links_2_files(self):
+		"""
+		Tests that broken links are detected.
+		"""
+		def reraise(e):
+			raise e
+
+		self.require_symlink()
+		self.make_links([
+			('A', 'DOES_NOT_EXIST'),
+		])
+		# `iter_tree_files` does not stat the link.
+		set(iter_tree_files(self.temp_dir, on_error=reraise))
+
+	def test_02_link_7_ignore_broken_links_1_entries(self):
+		"""
+		Tests that broken links are ignored.
+		"""
+		self.require_symlink()
+		self.make_links([
+			('A', 'DOES_NOT_EXIST'),
+		])
+		results = get_paths_from_entries(iter_tree_entries(self.temp_dir))
+		self.assertEqual(results, set())
+
+	def test_02_link_7_ignore_broken_links_2_files(self):
 		"""
 		Tests that broken links are ignored.
 		"""
@@ -351,7 +551,42 @@ class IterTreeTest(unittest.TestCase):
 		results = set(iter_tree_files(self.temp_dir))
 		self.assertEqual(results, set())
 
-	def test_02_link_8_no_follow_links(self):
+	def test_02_link_8_no_follow_links_1_entries(self):
+		"""
+		Tests to make sure directory links can be ignored.
+		"""
+		self.require_symlink()
+		self.make_dirs([
+			'Dir',
+		])
+		self.make_files([
+			'A',
+			'B',
+			'Dir/C',
+			'Dir/D',
+		])
+		self.make_links([
+			('Ax', 'A'),
+			('Bx', 'B'),
+			('Dir/Cx', 'Dir/C'),
+			('Dir/Dx', 'Dir/D'),
+			('DirX', 'Dir'),
+		])
+		results = get_paths_from_entries(iter_tree_entries(self.temp_dir, follow_links=False))
+		self.assertEqual(results, set(map(ospath, [
+			'A',
+			'Ax',
+			'B',
+			'Bx',
+			'Dir',
+			'Dir/C',
+			'Dir/Cx',
+			'Dir/D',
+			'Dir/Dx',
+			'DirX',
+		])))
+
+	def test_02_link_8_no_follow_links_2_files(self):
 		"""
 		Tests to make sure directory links can be ignored.
 		"""
@@ -383,36 +618,6 @@ class IterTreeTest(unittest.TestCase):
 			'Dir/D',
 			'Dir/Dx',
 			'DirX',
-		])))
-
-	def test_03_entries(self):
-		"""
-		Tests to make sure all files are found.
-		"""
-		self.make_dirs([
-			'Empty',
-			'Dir',
-			'Dir/Inner',
-		])
-		self.make_files([
-			'a',
-			'b',
-			'Dir/c',
-			'Dir/d',
-			'Dir/Inner/e',
-			'Dir/Inner/f',
-		])
-		results = {entry.path for entry in iter_tree_entries(self.temp_dir)}
-		self.assertEqual(results, set(map(ospath, [
-			'a',
-			'b',
-			'Dir',
-			'Dir/c',
-			'Dir/d',
-			'Dir/Inner',
-			'Dir/Inner/e',
-			'Dir/Inner/f',
-			'Empty',
 		])))
 
 
